requirements.txt
IMDbPY == 2022.7.9
requests == 2.31.0
bs4 == 0.0.2
flask == 3.0.3
gunicorn==22.0.0
Werkzeug==3.0.1